skip cleanup of failed archive

+++ CompressFilesToBlob.ps1
@@ -118,6 +118,8 @@ function IntegrityCheckFull {
         [string] $sourcePath
     )
 
+    Write-Verbose "Performing full integrity check of $filePath against $sourcePath..."
+
     # start a timer
     $stopwatch = [System.Diagnostics.Stopwatch]::StartNew()
 
@@ -126,7 +128,7 @@ function IntegrityCheckFull {
     $zipCRC = @{ }
     Write-Debug "Loading CRC from $filePath..."
     $params = @('l', '-slt', $filePath)
-    & $zipExe $params | ForEach-Object {
+    & $script:zipExe $params | ForEach-Object {
         if ($_.StartsWith('Path')) {
             $currentPath = $_.Substring($_.IndexOf('=') + 2)
         }
@@ -142,7 +144,7 @@ function IntegrityCheckFull {
     $errorCount = 0
     Write-Debug "Checking CRC from $sourcePath..."
     $params = @('h', $sourcePath)
-    & $zipExe $params | Select-Object -Skip 8 | ForEach-Object {
+    & $script:zipExe $params | Select-Object -Skip 8 | ForEach-Object {
         if (-not $endReached) {
             $crc = $_.Substring(0, 8)
             $path = $_.Substring(24)
@@ -175,10 +177,11 @@ function IntegrityCheckFull {
 
     if ($errorCount -gt 0) {
         Write-Warning "$errorCount error(s) detected. $itemsChecked items checked. Please check issues before continuing. ($($stopwatch.Elapsed))"
+        return $false
     }
-    else {
-        Write-Output "$filePath full integrity check completed successfully. $itemsChecked items checked. ($($stopwatch.Elapsed))"
-    }
+
+    Write-Information "$filePath full integrity check completed successfully. $itemsChecked items checked. ($($stopwatch.Elapsed))"
+    return $true
 }
 
 #####################################################################
@@ -190,15 +193,19 @@ function IntegrityCheckSimple {
         [string] $filePath
     )
 
+    Write-Verbose "Performing simple integrity check of $filePath..."
+    
     # start a timer
     $stopwatch = [System.Diagnostics.Stopwatch]::StartNew()
 
     $params = @('t', $filePath)
-    & $zipExe $params
+    & $script:zipExe $params
     if (-not $?) {
-        throw "Error testing archive - $filePath" 
+        return $false
     }   
-    Write-Output "$filePath simple integrity check completed successfully. ($($stopwatch.elapsed))"
+
+    Write-Information "$filePath simple integrity check completed successfully. ($($stopwatch.elapsed))"
+    return $true
 }
 
 #####################################################################
@@ -240,26 +247,34 @@ function CompressPathToBlob {
     $params += $sourcePath
     
     Write-Verbose "Archiving $sourcePath to $archivePath..."
-    & $zipExe $params
+    & $script:zipExe $params 
     if (-not $?) {
         Write-Error "Error creating archive: $archivePath from: $sourcePath"
         throw
     }
-    Write-Output "$archivePath created. ($($stopwatch.Elapsed))"
+    Write-Information "$archivePath created. ($($stopwatch.Elapsed))"
 
     # check compressed file
+    $result = $true
     if ($script:IntegrityCheck -eq 'None') {
         # skip the integrity check
     }
     elseif ($script:IntegrityCheck -eq 'Simple') {
-        IntegrityCheckSimple -filePath $archivePath
+        $result = IntegrityCheckSimple -filePath $archivePath
     }
     elseif ($script:IntegrityCheck -eq 'Full') {
-        IntegrityCheckFull -filePath $archivePath -sourcePath $sourcePath
+        $result = IntegrityCheckFull -filePath $archivePath -sourcePath $sourcePath
+
     } else {
+        $result = $false
         Write-Error "-IntegrityCheck $IntegrityCheck invaild. No integrity check performed." 
     }
 
+    if (-not $result) {
+        Write-Warning "ERROR -- found when comparing $archivePath to $sourcePath"
+    }
+    
+    $script:fileContinue = $result
 }
 
 #####################################################################
@@ -290,13 +305,17 @@ function CopyFileToContainer {
         $params += ("--block-blob-tier=$script:BlobTier")
     }
 
-    & $azCopyExe $params
+    & $script:azCopyExe $params
     if (-not $?) {
-        Write-Error "Error uploading file: $filePath to $containerURI"
-        throw
+        Write-Error "ERROR Context "ContextName" {
+            It "ItName" {
+                Assertion
+            }
+        } uploading file: $filePath to $containerURI"
+        $script:fileContinue = $false
     }
 
-    Write-Output "$($uri.Host)$($uri.LocalPath)/$($path.Name) copy complete. ($($stopwatch.Elapsed))"
+    Write-Information "$($uri.Host)$($uri.LocalPath)/$($path.Name) copy complete. ($($stopwatch.Elapsed))"
 }
 
 #####################################################################
@@ -354,11 +373,11 @@ Function CleanUpSource {
     # clean up source files
     if ($CleanUpDir -eq 'Delete') {
         Remove-Item -Path $sourcePath -Recurse -Force
-        Write-Output "$sourcePath deleted ($($stopwatch.Elapsed))"
+        Write-Information "$sourcePath deleted ($($stopwatch.Elapsed))"
 
     } elseif ($CleanUpDir) {
         Move-Item -Path $sourcePath -Destination $CleanUpDir
-        Write-Output "$sourcePath moved to $CleanupDir ($($stopwatch.Elapsed))"
+        Write-Information "$sourcePath moved to $CleanupDir ($($stopwatch.Elapsed))"
     }
 
 }
@@ -415,17 +434,9 @@ if ($UseManagedIdentity) {
         throw "Please login (Connect-AzAccount) and set the proper subscription context before proceeding."
     }    
     $environment = Get-AzEnvironment -Name $context.Environment
-
-    # login to azcopy
-    # $params = @('login', '--identity', "--aad-endpoint", $environment.ActiveDirectoryAuthority)
-    # & $azcopyExe $params
-    # if (-not $?) {
-    #     throw "Unable to login to azcopy using: $azcopyExe - $($params -join ' ')"
-    # }
 }
 
 if ($PSCmdlet.ParameterSetName -eq 'StorageAccount') { 
-    Write-Progress -Activity "Checking environment..." -Status "validating Storage Account"
     try {
         $result = Get-AzContext -ErrorAction Stop
         if (-not $result.Environment) {
@@ -494,8 +505,13 @@ if ($SeparateEachDirectory) {
     $sourcePaths = $SourceFilePath
 }
 
-$archiveCount = 0
+# remove the progress bar
+Write-Progress  -Activity "Checking environment..." -Completed
+
+$archiveSuccesses = @()
+$archiveFailures = @()
 foreach ($sourcePath in $sourcePaths) {
+    $fileContinue = $true
 
     if ($BlobName) {
         # only one large blob being created
@@ -533,25 +549,19 @@ foreach ($sourcePath in $sourcePaths) {
     Write-Output ''
 
     CompressPathToBlob -SourcePath $sourcePath -archivePath $archivePath
+    if (-not $fileContinue) {
+        $archiveFailures += $sourcePath
+        Write-Warning "ERROR -- $sourcePath failed compression to $archivePath. Please check errors and try again."
+        continue        
+    }
+
     CopyFileToContainer -filePath $archivePath -ContainerURI $ContainerURI
-
-    # if ($PSCmdlet.ParameterSetName -eq 'StorageAccount') { 
-    #     # verify the file & blob size created
-    #     $stopwatch = [System.Diagnostics.Stopwatch]::StartNew()
-    #     $archiveFile = Get-ChildItem $archivePath
-    #     $blob = Get-AzStorageBlob -Context $storageAccount.Context -Container $ContainerName -Blob $archiveFile.Name
-    #     if ($archiveFile.Length -ne $blob.Length) {
-    #         Write-Error "$($archiveFile.Name) size ($($archiveFile.Length)) does not match $($blob.Name) Blob size ($($blob.Length))"
-    #         throw
-    #     }
-    #     Write-Output "$($archiveFile.Name) blob copy verified successfully. ($($stopwatch.Elapsed))"
-
-    #     if ($BlobTier) {
-    #         $blob.ICloudBlob.SetStandardBlobTier($BlobTier)
-    #         Write-Output "$containerURI/$($blob.Name) tier set to $BlobTier"
-    #     }
-    # }
- 
+    if (-not $fileContinue) {
+        $archiveFailures += $sourcePath
+        Write-Warning "ERROR -- $archivePath failed copy to $ContainerURI. Please check errors and try again."
+        continue
+    }
+    
     # clean up zip file
     Remove-Item -Path $archivePath -Force
 
@@ -564,8 +574,24 @@ foreach ($sourcePath in $sourcePaths) {
     Write-Output "==================== $(Split-Path $archivePath -Leaf) complete. $(Get-Date) ===================="
     Write-Output ''
     
-    $archiveCount++
+    $archiveSuccesses += $sourcePath
 }
 
-Write-Output "$archiveCount archives copied."
+if ($archiveSuccesses) {
+    Write-Output ''
+    Write-Output '===== SUCCESSFULLY PROCESSED ====='
+    Write-Output $archiveSuccesses
+}
+
+if ($archiveFailures) {
+    Write-Output ""
+    Write-Output "===== FAILED ====="
+    Write-Output $archiveFailures
+}
+
+Write-Output ""
+Write-Output "===== FINAL STATS ====="
+Write-Output "$($archiveSuccesses.Count) succeeded"
+Write-Output "$($achiveFailures.count) failed"
+Write-Output ""
 Write-Output "Script Complete. $(Get-Date)"
